requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiolimiter==1.0.0
aioredis==1.3.1
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
Django==3.1.12
djongo==1.3.6
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-login==1.9.1
fastapi-pagination==0.12.8
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hiredis==2.2.3
hpack==4.0.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
openai==0.27.8
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pendulum==2.1.2
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pymongo==3.12.3
pypi.testpkg==8.2
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.6
python-telegram-bot==20.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.5.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.2.4
starlette==0.27.0
tiktoken==0.4.0
tornado==6.3.2
tortoise-orm==0.20.0
tqdm==4.65.0
trio==0.21.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
